refactor(generator): rename generator package to be generate
